Version 1.0.5
It works... barely. This is the bare bones of the application. It works, however it is missing a lot of funcitonality i want to add later on. for example, I want to make it so if the user doesnt click anywhere on the screen and tries to run the program it will generate random starting alive squares on the page.